build: add ability to build vxworks in main branch
This patch updates the source code tree to build in the main branch.
This takes some of the changes in the vxworks_layer branch, but making
the vxworks Makefile utilizing the scripts for building that are also
used by the Android build system.  This ensures that generated files are
consistent between all builds.
