+++ build-sys/vxworks-osal/Makefile
@@ -1,0 +1,20 @@
+#
+# Copyright (C) 2018 Wind River Systems, Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at:
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#
+
+PRE_BUILD_DIRS = lib
+
+BUILD_DIRS = lib
+BUILD_USER_DIRS = lib
+
+include $(WIND_KRNL_MK)/rules.layers.mk
