+++ src/os_posix_private.h
@@ -23,7 +23,6 @@
 
 /* includes for #defined functions */
 #include <ctype.h>
-#include <dlfcn.h>
 #include <errno.h>
 #include <limits.h> /* for PATH_MAX */
 #include <pthread.h>
@@ -33,22 +32,13 @@
 #include <sys/time.h> /* for gettimeofday */
 #include <unistd.h> /* for sleep */
 #include <time.h>   /* for nanosleep */
+#include <uuid/uuid.h>   /* libuuid functions + uuid_t */
 
-#ifndef _WRS_KERNEL
-#include <sys/utsname.h> /* for struct utsname */
-#endif
-
-#ifndef _WRS_KERNEL
-#	include <uuid/uuid.h>   /* osals uuid header, not system, for libuuid functions + uuid_t */
-#endif
-
-#ifndef _WRS_KERNEL
 /** @brief Maximum length of a host name on POSIX systems */
-#ifndef _POSIX_HOST_NAME_MAX
+#if !defined( _POSIX_HOST_NAME_MAX )
 #	define _POSIX_HOST_NAME_MAX 64
-#endif
+#endif /* if !defined( _POSIX_HOST_NAME_MAX ) */
 #define OS_HOST_MAX_LEN _POSIX_HOST_NAME_MAX
-#endif
 
 #include <arpa/inet.h>  /* for inet_ntoa, inet_htons */
 #include <dirent.h>     /* for DIR*, opendir, closedir */
@@ -60,7 +50,7 @@
 #include <signal.h>     /* for siginfo_t */
 #include <sys/socket.h> /* for struct addrinfo */
 
-#if defined(__linux__) || defined (_WRS_KERNEL)
+#if defined(__linux__) || defined (__VXWORKS__)
 #	include <net/if.h>    /* for struct ifconf */
 #endif
 #include <ifaddrs.h>
