+++ src/os_vxworks.c
@@ -1,8 +1,8 @@
 /**
- * @file  os_vxworks.c
- * @brief VxWorks OS adaptation layer
+ * @file
+ * @brief VxWorks operating system abstraction layer
  *
- * @copyright Copyright (C) 2016-2017 Wind River Systems, Inc. All Rights Reserved.
+ * @copyright Copyright (C) 2016-2018 Wind River Systems, Inc. All Rights Reserved.
  *
  * @license Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -10,25 +10,156 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
- * OR CONDITIONS OF ANY KIND, either express or implied."
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied."
  */
 
-#ifdef _WRS_KERNEL
-#include <vxWorks.h>
 #include "os.h"
-#include <uuid.h>
-#include <semLib.h>
-#include <kernelLib.h>
-#include <version.h>
-#include <string.h>
-#include <agent_config.h>
-#include <taskLib.h>
-#include <types.h>
 
-extern char *sysModel (void);
-extern STATUS sysToMonitor (int startType);
+#include <ioLib.h>
+#include <pthread.h>
+#include <semLib.h>
+#include <stdlib.h> /* for EXIT_SUCCESS */
+#include <string.h>
+#include <taskLib.h>
+#include <rtpLib.h>
+#include <wait.h>
+#include <version.h>
+#include <vxWorks.h>
+
+#if defined(_WRS_KERNEL)
+#include <sysLib.h>
+#include <vsbConfig.h>
+#include <bootLib.h>
+#include <shellLib.h>
+#ifdef _WRS_CONFIG_SYS_PWR_OFF
+#include <powerOffLib.h>
+#endif /* _WRS_CONFIG_SYS_PWR_OFF */
+extern BOOT_PARAMS sysBootParams;
+#endif /* _WRS_KERNEL */
+
+#define VX_RW_SEM_MAX_READERS (255)
+
+#if defined(_WRS_KERNEL)
+extern int control_main ( int argc, char* argv[] );
+extern int iot_update_main ( int argc, char* argv[] );
+#endif /* if(defined(_WRS_KERNEL) */
+
+/* directory operations */
+os_uint64_t os_directory_free_space( const char* path )
+{
+	return 0;
+}
+
+/** @todo fix later once it is needed in vxWorks */
+os_status_t os_directory_delete(
+	const char *path, const char *regex, os_bool_t recursive )
+{
+	os_status_t result = OS_STATUS_FAILURE;
+
+	if (regex == NULL && rmdir (path) == OK)
+		result = OS_STATUS_SUCCESS;
+
+	return result;
+}
+
+/* file operations */
+os_status_t os_file_chown(
+	const char *path,
+	const char *user )
+{
+	return OS_STATUS_FAILURE;
+}
+
+
+/* process functions */
+os_status_t os_path_executable(
+	char *path,
+	size_t size )
+{
+	os_status_t result = OS_STATUS_BAD_PARAMETER;
+	if ( path )
+	{
+		const char *task_name = taskName(taskIdSelf());
+		result = OS_STATUS_FAILURE;
+		if (task_name && size > 2u)
+		{
+			path[0] = OS_DIR_SEP;
+			strncpy(&path[1], task_name, size - 1u);
+			result = OS_STATUS_SUCCESS;
+		}
+	}
+	return result;
+}
+
+os_status_t os_process_cleanup( void )
+{
+	return OS_STATUS_FAILURE;
+}
+
+/* service functions */
+os_status_t os_service_run(
+	const char *id,
+	os_service_main_t service_function,
+	int argc,
+	char *argv[],
+	int remove_argc,
+	const char *remove_argv[],
+	os_sighandler_t UNUSED(handler),
+	const char *UNUSED(logdir) )
+{
+	os_status_t result = OS_STATUS_BAD_PARAMETER;
+	if ( id && service_function )
+	{
+		int i;
+		char** good_argv = (char**)malloc( (unsigned long)argc * sizeof( char* ) );
+		result = OS_STATUS_FAILURE;
+		/* remove bad arguments */
+		if ( good_argv )
+		{
+			int good_argc = 0u;
+			for ( i = 0; i < argc; ++i )
+			{
+				int is_good_arg = 1; /* true */
+				if ( remove_argc > 0 && remove_argv && argv[i] )
+				{
+					int j;
+					for ( j = 0;
+						j < remove_argc && is_good_arg; ++j )
+					{
+						if ( remove_argv[j] )
+						{
+							size_t arg_len =
+								strlen( remove_argv[j] );
+							is_good_arg = strncmp( argv[i],
+								remove_argv[j], arg_len );
+						}
+					}
+				}
+
+				if ( is_good_arg )
+				{
+					good_argv[good_argc] = argv[i];
+					++good_argc;
+				}
+			}
+
+			if ( ( *service_function )( good_argc, good_argv ) == EXIT_SUCCESS )
+				result = OS_STATUS_SUCCESS;
+
+			free( good_argv );
+		}
+	}
+	return result;
+}
+
+/** @todo fix later once it needs in vxWorks */
+os_status_t os_stream_echo_set(
+	os_file_t stream, os_bool_t enable )
+{
+	return OS_STATUS_FAILURE;
+}
 
 os_status_t os_system_info(
 	os_system_info_t *sys_info )
@@ -42,105 +173,29 @@ os_status_t os_system_info(
 		* VxWorks, provided by Wind River Systems.
 		*/
 
-		strncpy( sys_info->vendor_name, "Wind River Systems", OS_SYSTEM_INFO_MAX_LEN );
-		strncpy( sys_info->system_name, "VxWorks", OS_SYSTEM_INFO_MAX_LEN );
-		strncpy( sys_info->system_version, runtimeVersion, OS_SYSTEM_INFO_MAX_LEN );
-		strncpy( sys_info->system_release, _WRS_CONFIG_CORE_KERNEL_VERSION, OS_SYSTEM_INFO_MAX_LEN );
-		strncpy( sys_info->system_platform, sysModel(), OS_SYSTEM_INFO_MAX_LEN );
+		strncpy( sys_info->vendor_name, "Wind River Systems",
+			OS_SYSTEM_INFO_MAX_LEN );
+		strncpy( sys_info->system_name, RUNTIME_NAME,
+			OS_SYSTEM_INFO_MAX_LEN );
+		strncpy( sys_info->system_version, VXWORKS_VERSION,
+			OS_SYSTEM_INFO_MAX_LEN );
+		strncpy( sys_info->system_release,
+			_WRS_CONFIG_CORE_KERNEL_VERSION,
+			OS_SYSTEM_INFO_MAX_LEN );
+		strncpy( sys_info->system_platform, sysModel(),
+			OS_SYSTEM_INFO_MAX_LEN );
+#if defined(_WRS_KERNEL)
+		strncpy( sys_info->host_name, sysBootParams.targetName,
+			OS_SYSTEM_INFO_MAX_LEN );
+#else /* if defined(_WRS_KERNEL) */
 		strncpy( sys_info->host_name, "", OS_SYSTEM_INFO_MAX_LEN );
+#endif /* else if defined(_WRS_KERNEL) */
 		sys_info->system_flags = 0;
 	}
 	return OS_STATUS_SUCCESS;
 }
 
-os_lib_handle os_library_open(
-	const char *path )
-{
-	return 0;
-}
-
-os_status_t os_library_close(
-	os_lib_handle lib )
-{
-	return OS_STATUS_FAILURE;
-}
-
-void *os_library_find(
-	os_lib_handle lib,
-	const char *function )
-{
-	return NULL;
-}
-
-os_uint64_t os_directory_free_space( const char* path )
-{
-	return 0;
-}
-
-os_status_t os_file_temp(
-	char *prototype,
-	size_t suffix_len)
-{
-	return OS_STATUS_FAILURE;
-}
-
-os_status_t os_process_cleanup( void )
-{
-	return 0;
-}
-
-os_status_t os_file_chown(
-	const char *path,
-	const char *user )
-{
-	return 0;
-}
-
-/** @todo fix later once it needs in vxWorks */
-os_status_t os_directory_delete(
-	const char *path, const char *regex, os_bool_t recursive )
-{
-	return 0;
-}
-
-/** @todo fix later once it needs in vxWorks */
-os_status_t os_stream_echo_set(
-	os_file_t stream, os_bool_t enable )
-{
-	return 0;
-}
-
-/* NOTE: All API library clients share the pthread_attr structure. */
-static pthread_attr_t pthread_attr;
-static pthread_attr_t *pPthread_attr = NULL;
-
-os_status_t os_thread_create(
-	os_thread_t *thread,
-	os_thread_main_t main,
-	void *arg )
-{
-	os_status_t result = OS_STATUS_BAD_PARAMETER;
-
-	if ( main )
-	{
-		result = OS_STATUS_FAILURE;
-
-		if ( NULL == pPthread_attr )
-		{
-			/* Should setup the structure the first time through */
-			if ( ( 0 == pthread_attr_init( &pthread_attr ) ) &&
-				( 0 == pthread_attr_setstacksize( &pthread_attr, get_hdc_agent_worker_thread_stack_size() ) ) )
-				pPthread_attr = &pthread_attr;
-			else
-			return result;
-		}
-
-		if ( pthread_create( thread, pPthread_attr, main, arg ) == 0 )
-			result = OS_STATUS_SUCCESS;
-	}
-	return result;
-}
-
+#if defined(OSAL_THREAD_SUPPORT) && OSAL_THREAD_SUPPORT
 os_status_t os_thread_rwlock_create(
 	os_thread_rwlock_t *lock )
 {
@@ -154,7 +209,6 @@ os_status_t os_thread_rwlock_create(
 		if ( SEM_ID_NULL != semId )
 		{
 			*lock = semId;
-			result = OS_STATUS_SUCCESS;
 		}
 	}
 
@@ -235,76 +289,167 @@ os_status_t os_thread_rwlock_destroy(
 
 	return result;
 }
+#endif /* if defined(OSAL_THREAD_SUPPORT) && OSAL_THREAD_SUPPORT */
+
+#if defined(_WRS_KERNEL)
+static void os_vxworks_reboot( void )
+{
+	/* Wait 5 seconds for messages to propagate */
+
+	sleep (5);
+
+	/* Force a cold reboot - We do not return */
+
+	sysToMonitor(2);
+}
+
+#if defined(_WRS_CONFIG_SYS_PWR_OFF)
+static void os_vxworks_shutdown( void )
+{
+	/* Wait 5 seconds for messages to propagate */
+
+	sleep (5);
+
+	powerOff();
+}
+#endif /* _WRS_CONFIG_SYS_PWR_OFF */
+
+static os_status_t os_vxworks_script(
+	char * script )
+{
+	os_status_t result = OS_STATUS_FAILURE;
+	char * shellTaskName;
+	int fd;
+
+	if ((fd = open(script, O_RDONLY, 0)) == ERROR) {
+		return OS_STATUS_FAILURE;
+	}
+
+	if (shellGenericInit("INTERPRETER=C", 0, NULL, &shellTaskName, FALSE,
+		FALSE, fd, STD_OUT, STD_ERR) == OK) {
+		result = OS_STATUS_SUCCESS;
+	}
+
+	do {
+		taskDelay(sysClkRateGet());
+	} while (taskNameToId(shellTaskName) != TASK_ID_ERROR);
+
+	close(fd);
+
+	return result;
+}
+#endif /* _WRS_KERNEL */
 
 os_status_t os_system_run(
 	const char *command,
 	int *exit_status,
-	os_file_t pipe_files )
+	os_bool_t privileged,
+	int priority,
+	size_t stack_size,
+	os_file_t pipe_files[2u] )
 {
+	static const char * argv[10];
+	int argc = 0;
+
+	/* set a default exit status */
+
+	if ( exit_status )
+		*exit_status = -1;
+
+	/* tokenize the command */
+	argv[argc] = strtok (command, " ");
+	while ((argv[argc] != NULL) && (++argc < 9)) {
+		argv[argc] = strtok (NULL, " ");
+	}
+	argv[9] = NULL;
+
 	/*
 	 * Go through list of supported commands
 	 */
+#if defined(_WRS_KERNEL)
+	if (strstr (argv[0], "iot-control") != NULL) {
+		if (taskSpawn ("tControl",
+			priority, /*deviceCloudPriorityGet(), */ 0,
+			stack_size, /* deviceCloudStackSizeGet(), */
+			(FUNCPTR) control_main,
+			argc, argv, 0, 0, 0, 0, 0, 0, 0, 0) == TASK_ID_ERROR) {
+			return OS_STATUS_FAILURE;
+		}
+	} else if (strstr (argv[0], "iot-update") != NULL) {
+		if (taskSpawn ("tUpdate",
+			priority, /* deviceCloudPriorityGet(), */ 0,
+			stack_size, /* deviceCloudStackSizeGet(), */
+			(FUNCPTR) iot_update_main,
+			argc, argv, 0, 0, 0, 0, 0, 0, 0, 0) == TASK_ID_ERROR) {
+			return OS_STATUS_FAILURE;
+		}
+	} else if (strncmp (argv[0], "sh", sizeof("sh")) == 0) {
+		if (os_vxworks_script(argv[1]) == OS_STATUS_FAILURE) {
+			return OS_STATUS_FAILURE;
+		}
+	} else {
+		if (os_vxworks_script(argv[0]) == OS_STATUS_FAILURE) {
+			return OS_STATUS_FAILURE;
+		}
+	}
+#else
+	if (rtpSpawn(argv[0], argv, NULL,
+		priority, /*deviceCloudPriorityGet(), */
+		stack_size, /* deviceCloudStackSizeGet(),*/
+		RTP_LOADED_WAIT, VX_FP_TASK) == RTP_ID_ERROR) {
+		return OS_STATUS_FAILURE;
+	}
+	else {
+	       printf("Invalid command:%s\n", command);
+	       return OS_STATUS_FAILURE;
+	}
+#endif /* defined(_WRS_KERNEL) */
 
-	if (strncmp (command, "reboot", 6) == 0)
-		sysToMonitor(2);
+	if ( exit_status )
+		*exit_status = 0;
 
-	*exit_status = 0;
 	return OS_STATUS_SUCCESS;
 }
 
 os_status_t os_system_run_wait(
 	const char *command,
 	int *exit_status,
+	os_bool_t privileged,
+	int priority,
+	size_t stack_size,
 	char *out_buf[2u],
 	size_t out_len[2u],
 	os_millisecond_t UNUSED(max_time_out) )
 {
 	os_file_t pipes[2u] = {NULL, NULL};
-	return os_system_run(command, exit_status, pipes, 0);
+	os_status_t result = os_system_run(command, exit_status, privileged,
+		priority, stack_size, pipes);
+	return result;
 }
 
-/* uuid support */
-os_status_t os_uuid_generate(
-	os_uuid_t *uuid )
+os_status_t os_system_shutdown(
+	os_bool_t reboot, unsigned int delay)
 {
-	uint32_t status;
-	os_status_t result = OS_STATUS_BAD_PARAMETER;
-
-	if ( uuid )
+	os_status_t result = OS_STATUS_FAILURE;
+#if defined(_WRS_KERNEL)
+	if (reboot != OS_FALSE)
 	{
-		uuid_create( uuid, &status );
-		if ( uuid_s_ok == status )
+		if (taskSpawn ("tReboot", 10, 0, 0x1000,
+			(FUNCPTR) os_vxworks_reboot,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0) != TASK_ID_ERROR) {
 			result = OS_STATUS_SUCCESS;
+		}
 	}
-
+#if defined(_WRS_CONFIG_SYS_PWR_OFF)
+	else
+	{
+		if (taskSpawn ("tShutdown", 10, 0, 0x1000,
+			(FUNCPTR) os_vxworks_shutdown,
+			0, 0, 0, 0, 0, 0, 0, 0, 0, 0) != TASK_ID_ERROR) {
+			result = OS_STATUS_SUCCESS;
+		}
+	}
+#endif /* if defined(_WRS_CONFIG_SYS_PWR_OFF) */
+#endif /* if defined(_WRS_KERNEL) */
 	return result;
 }
-
-os_status_t os_uuid_to_string_lower(
-	os_uuid_t *uuid,
-	char *dest,
-	size_t len )
-{
-	uint32_t status;
-	os_status_t result = OS_STATUS_BAD_PARAMETER;
-
-	if ( uuid && dest && ( len >= 37 ) )
-        {
-		uuid_to_string( uuid, dest, &status );
-		if ( uuid_s_ok == status )
-			result = OS_STATUS_SUCCESS;
-	}
-	return result;
-}
-
-os_uint32_t os_system_pid( void )
-{
-	/*
-	* VxWorks TASK_ID is a 64-bit pointer on a 64-bit host... We need to get the
-	* ID and convert it to a non-pointer and then cast to the return type to
-	* avoid a compiler warning.
- 	*/
-	return (os_uint32_t) (ULONG) taskIdSelf();
-}
-
-#endif /* _WRS_KERNEL */
