+++ src/uuid/randutils.c
@@ -13,7 +13,9 @@
 #include <string.h>
 #include <sys/time.h>
 
+#if !defined(__VXWORKS__)
 #include <sys/syscall.h>
+#endif /* !defined(__VXWORKS__) */
 
 #include "randutils.h"
 
@@ -28,6 +30,7 @@
 THREAD_LOCAL unsigned short ul_jrand_seed[3];
 #endif
 
+#if !defined(__VXWORKS__)
 int random_get_fd(void)
 {
 	int i, fd;
@@ -55,6 +58,7 @@ int random_get_fd(void)
 		rand();
 	return fd;
 }
+#endif /* if !defined(__VXWORKS__) */
 
 
 /*
@@ -65,13 +69,22 @@ int random_get_fd(void)
 void random_get_bytes(void *buf, size_t nbytes)
 {
 	size_t i, n = nbytes;
+#if !defined(__VXWORKS__)
 	int fd = random_get_fd();
+#endif /* if !defined(__VXWORKS__) */
 	int lose_counter = 0;
 	unsigned char *cp = (unsigned char *) buf;
 
+#if !defined(__VXWORKS__)
 	if (fd >= 0) {
+#endif /* if !defined(__VXWORKS__) */
 		while (n > 0) {
+#if defined(__VXWORKS__)
+			ssize_t x = (ssize_t)1L;
+			*cp = (unsigned char)(rand() & 0xff);
+#else /* if defined(__VXWORKS__) */
 			ssize_t x = read(fd, cp, n);
+#endif /* else if defined(__VXWORKS__) */
 			if (x <= 0) {
 				if (lose_counter++ > 16)
 					break;
@@ -82,8 +95,10 @@ void random_get_bytes(void *buf, size_t nbytes)
 			lose_counter = 0;
 		}
 
+#if !defined(__VXWORKS__)
 		close(fd);
 	}
+#endif /* if !defined(__VXWORKS__) */
 
 	/*
 	 * We do this all the time, but this is the only source of
